Move ResourceType into Resources
