+++ sdk/cosmos/src/clients/attachment_client.rs
@@ -1,5 +1,6 @@
 use crate::requests;
-use crate::{ReadonlyString, ResourceType};
+use crate::resources::ResourceType;
+use crate::ReadonlyString;
 use azure_core::{HttpClient, No};
 
 use super::*;
