+++ sdk/cosmos/src/clients/document_client.rs
@@ -1,6 +1,6 @@
 use super::{AttachmentClient, CollectionClient, CosmosClient, DatabaseClient};
-use crate::requests;
-use crate::{PartitionKeys, ReadonlyString, ResourceType};
+use crate::resources::ResourceType;
+use crate::{requests, PartitionKeys, ReadonlyString};
 use azure_core::HttpClient;
 
 #[derive(Debug, Clone)]
