+++ sdk/cosmos/src/clients/stored_procedure_client.rs
@@ -1,6 +1,6 @@
 use super::*;
-use crate::requests;
-use crate::{ReadonlyString, ResourceType};
+use crate::resources::ResourceType;
+use crate::{requests, ReadonlyString};
 use azure_core::{HttpClient, No};
 
 #[derive(Debug, Clone)]
