+++ sdk/cosmos/src/clients/user_client.rs
@@ -1,5 +1,6 @@
 use super::*;
-use crate::{requests, ReadonlyString, ResourceType};
+use crate::resources::ResourceType;
+use crate::{requests, ReadonlyString};
 use azure_core::{HttpClient, No};
 
 #[derive(Debug, Clone)]
