+++ sdk/cosmos/src/requests/create_collection_builder.rs
@@ -1,7 +1,7 @@
 use crate::prelude::*;
 use crate::resources::collection::{Collection, CollectionName, IndexingPolicy, PartitionKey};
+use crate::resources::ResourceType;
 use crate::responses::CreateCollectionResponse;
-use crate::ResourceType;
 use azure_core::prelude::*;
 use azure_core::{No, ToAssign, Yes};
 use http::StatusCode;
