+++ sdk/cosmos/src/requests/get_partition_key_ranges_builder.rs
@@ -1,6 +1,6 @@
 use crate::prelude::*;
+use crate::resources::ResourceType;
 use crate::responses::GetPartitionKeyRangesResponse;
-use crate::ResourceType;
 use azure_core::prelude::*;
 use chrono::{DateTime, Utc};
 use http::StatusCode;
