+++ sdk/cosmos/src/requests/list_documents_builder.rs
@@ -1,6 +1,6 @@
 use crate::prelude::*;
+use crate::resources::ResourceType;
 use crate::responses::ListDocumentsResponse;
-use crate::ResourceType;
 use azure_core::prelude::*;
 use futures::stream::{unfold, Stream};
 use http::StatusCode;
