+++ sdk/cosmos/src/requests/list_triggers_builder.rs
@@ -1,6 +1,6 @@
 use crate::prelude::*;
+use crate::resources::ResourceType;
 use crate::responses::ListTriggersResponse;
-use crate::ResourceType;
 use azure_core::prelude::*;
 use futures::stream::{unfold, Stream};
 use http::StatusCode;
