+++ sdk/cosmos/src/requests/query_documents_builder.rs
@@ -1,6 +1,7 @@
 use crate::prelude::*;
+use crate::resources::ResourceType;
 use crate::responses::QueryDocumentsResponse;
-use crate::{Query, ResourceType};
+use crate::Query;
 use azure_core::prelude::*;
 use azure_core::{No, ToAssign, Yes};
 use chrono::{DateTime, Utc};
